tests/wpt/tests/workers/interfaces/WorkerUtils/importScripts/011.html

<!doctype html>
<title>importScripts(null)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('011.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 'null');
    this.done();
  });
  worker.onerror = this.step_func(function(e) {
    assert_unreached(e.message);
  });
});
</script>
